pospondokprogrammer

+++ app/Http/Controllers/Api/PembelianController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
+use App\Barang;
+use App\Http\Controllers\Api\BaseController;
 use App\Pembelian;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
-class PembelianController extends Controller
+class PembelianController extends BaseController
 {
     /**
      * Display a listing of the resource.
@@ -18,10 +19,11 @@ class PembelianController extends Controller
     {
         $pembelian = Pembelian::get();
 
-        if (empty($pembelian)) {
-            return $this->responseError('Pembelian Kosong', 403);
+        if ($pembelian == []) {
+            return $this->responseError('Pengeluaran belum ada');
+        } else {
+            return $this->responseOk($pembelian);
         }
-        return $this->responseOk($pembelian, 200, 'Sukses Liat Data Pembelian');
     }
 
     /**
@@ -44,20 +46,29 @@ class PembelianController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'supplier_id' => 'required|integer',
-            'barang' => 'required|string',
+            'barang_id' => 'required|integer',
+            'jumlah' => 'required|integer',
+            'total_biaya' => 'required|integer',
         ]);
 
         if ($validator->fails()) {
-            return $this->responseError('Gagal Buat Pembelian', 422, $validator->errors());
+            return $this->responseError('Pembelian gagal ditambah', 422, $validator->errors());
         }
 
         $params = [
             'supplier_id' => $request->supplier_id,
-            'barang' => $request->barang,
+            'barang_id' => $request->barang_id,
+            'jumlah' => $request->jumlah,
+            'total_biaya' => $request->total_biaya,
         ];
 
         $pembelian = Pembelian::create($params);
-        return $this->responseOk($pembelian, 200, 'Sukses Buat Pembelian');
+        $barang = Barang::find($pembelian->barang_id);
+        $data['stok'] = $barang->stok + $pembelian->jumlah;
+        $data['harga_beli'] = $pembelian->total_biaya / $pembelian->jumlah;
+        $barang->update($data);
+
+        return $this->responseOk($pembelian->load('supplier', 'barang'), 201, 'Pembelian berhasil ditambah');
     }
 
     /**
@@ -92,23 +103,50 @@ class PembelianController extends Controller
     public function update(Request $request, $id)
     {
         $validator = Validator::make($request->all(), [
-            'supplier_id' => 'string',
-            'barang' => 'string',
+            'supplier_id' => 'integer',
+            'barang_id' => 'integer',
+            'jumlah' => 'integer',
+            'total_biaya' => 'integer',
         ]);
 
         if ($validator->fails()) {
-            return $this->responseError('Gagal Ubah Pembelian', 422, $validator->errors());
+            return $this->responseError('Pembelian gagal diubah', 422, $validator->errors());
         }
 
         $pembelian = Pembelian::find($id);
+        $jumlahOld = $pembelian->jumlah;
+        $barang_idOld = $pembelian->barang_id;
+        $barangOld = Barang::find($barang_idOld);
+        $stokOld = $barangOld->stok;
 
         $params = [
             'supplier_id' => $request->supplier_id ?? $pembelian->supplier_id,
-            'barang' => $request->barang ?? $pembelian->barang,
+            'barang_id' => $request->barang_id ?? $pembelian->barang_id,
+            'jumlah' => $request->jumlah ?? $pembelian->jumlah,
+            'total_biaya' => $request->total_biaya ?? $pembelian->total_biaya,
         ];
 
+        if ($request->jumlah > $stokOld) {
+            return $this->responseError('Pembelian gagal diubah karena barang sudah terjual');
+        }
+
         $pembelian->update($params);
-        return $this->responseOk($pembelian, 200, 'Sukses Ubah Pembelian');
+        $barang = Barang::find($pembelian->barang_id);
+
+        if ($barang_idOld !== $pembelian->barang_id) {
+            $dataOld['stok'] = $barangOld->stok - $jumlahOld;
+            $barangOld->update($dataOld);
+
+            $data['stok'] =  $barang->stok + $pembelian->jumlah;
+            $data['harga_beli'] = $pembelian->total_biaya / $pembelian->jumlah;
+            $barang->update($data);
+        } else {
+            $data['stok'] = $barang->stok - $jumlahOld + $pembelian->jumlah;
+            $data['harga_beli'] = $pembelian->total_biaya / $pembelian->jumlah;
+            $barang->update($data);
+        }
+
+        return $this->responseOk($pembelian->load('supplier', 'barang'), 200, 'Pembelian berhasil diubah');
     }
 
     /**
@@ -120,8 +158,15 @@ class PembelianController extends Controller
     public function destroy($id)
     {
         $pembelian = Pembelian::find($id);
-        $pembelian->delete();
+        $barang = Barang::find($pembelian->barang_id);
+        if ($barang->stok < $pembelian->jumlah) {
+            return $this->responseError('Pembelian gagal dihapus karena barang sudah terjual');
+        } else {
+            $data['stok'] = $barang->stok - $pembelian->jumlah;
+            $barang->update($data);
+            $pembelian->delete();
+        }
 
-        return $this->responseOk(null, 200, 'Sukses Hapus Pembelian');
+        return $this->responseOk(null, 200, 'Pembelian berhasil dihapus');
     }
 }
